Split worker btwn manager & child process
